Many changes in the Data class.
Now the whole working principle is much easier to understand, and both
instrument should work correctly. Note that there is a weakness point in
the fund_instrument function, since it is not so clear how to recognize
the instrument at all. I will try to implement a solution in the
following days.